Move renamed CopyJson implementations to a different namespace for better cohesion
