+++ app/Event/EventServiceProvider.php
@@ -3,7 +3,7 @@
 namespace CultuurNet\UDB3\SearchService\Event;
 
 use CultuurNet\UDB3\Search\ElasticSearch\ElasticSearchDocumentRepository;
-use CultuurNet\UDB3\Search\ElasticSearch\Event\EventJsonDocumentTransformer;
+use CultuurNet\UDB3\Search\ElasticSearch\JsonDocument\EventJsonDocumentTransformer;
 use CultuurNet\UDB3\Search\ElasticSearch\PathEndIdUrlParser;
 use CultuurNet\UDB3\Search\Event\EventSearchProjector;
 use CultuurNet\UDB3\Search\JsonDocument\TransformingJsonDocumentIndexService;
